requirements.txt
tensorflow
pandas
tweet-preprocessor
spacy
pandarallel
bert-tensorflow
tensorflow_hub
PyYAML